requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
geopy==2.3.0
twilio==8.10.0
python-dotenv==1.0.0
sqlite3
datetime
geocoder==1.38.1
phonenumbers==8.13.25
email-validator==2.1.0
schedule==1.2.0